Created php file
This file will allow an email to be sent to the union's desired email. It is a lttle work around from a more complex version. The only down sides are that if the form gets more complex this might not be the best way forward. Also emails at firs may be sent to the spam folder.